# packages/contracts/openapi/bifrost.v1.yaml
openapi: 3.1.0
info:
  title: Bifrost API
  version: 0.1.0
paths:
...
  /api/v1/admin/roles:
    get:
      tags: [AdminRoles]
      operationId: listAdminRoles
      summary: List roles.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
    post:
      tags: [AdminRoles]
      operationId: createAdminRole
      summary: Create a role.
      security:
        - BearerAuth: []
      responses:
        "201":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/roles/{roleId}:
    patch:
      tags: [AdminRoles]
      operationId: updateAdminRole
      summary: Update role display fields.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoleId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/roles/{roleId}/services:
    put:
      tags: [AdminPolicy]
      operationId: replaceRoleServices
      summary: Replace services granted to a role.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RoleId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/services:
    get:
      tags: [AdminServices]
      operationId: listAdminServices
      summary: List service catalog entries.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
        - name: group
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
    post:
      tags: [AdminServices]
      operationId: createAdminService
      summary: Create a service catalog entry.
      security:
        - BearerAuth: []
      responses:
        "201":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/services/{serviceId}:
    get:
      tags: [AdminServices]
      operationId: getAdminService
      summary: Return one service catalog entry.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
    patch:
      tags: [AdminServices]
      operationId: updateAdminService
      summary: Update service catalog fields.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/services/{serviceId}/status:
    post:
      tags: [AdminServices]
      operationId: setAdminServiceStatus
      summary: Enable or disable a service.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/devices:
    get:
      tags: [AdminDevices]
      operationId: listAdminDevices
      summary: List devices.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Keyword"
        - $ref: "#/components/parameters/Status"
        - name: userId
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/devices/{deviceId}:
    get:
      tags: [AdminDevices]
      operationId: getAdminDevice
      summary: Return one device.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/admin/devices/{deviceId}/status:
    post:
      tags: [AdminDevices]
      operationId: setAdminDeviceStatus
      summary: Trust or disable a device.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/DeviceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
...
  /api/v1/client/services:
    get:
      tags: [ClientServices]
      operationId: listClientServices
      summary: List services available to the active user and device.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/Keyword"
        - name: group
          in: query
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
  /api/v1/client/services/{serviceId}:
    get:
      tags: [ClientServices]
      operationId: getClientService
      summary: Return a client-visible service.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ServiceId"
      responses:
        "200":
          $ref: "#/components/responses/ApiSuccess"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: BifrostAccessToken
    ServiceAccessCookie:
      type: apiKey
      in: cookie
      name: bifrost_access_ticket
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        default: 20
    Keyword:
      name: keyword
      in: query
      schema:
        type: string
    Status:
      name: status
      in: query
      schema:
        type: string
...
    RoleId:
      name: roleId
      in: path
      required: true
      schema:
        type: string
    ServiceId:
      name: serviceId
      in: path
      required: true
      schema:
        type: string
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ApiSuccess:
      description: Unified Bifrost success envelope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiSuccessEnvelope"
...
  schemas:
...
    Pagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    ApiMeta:
      type: object
      additionalProperties: false
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
        timestamp:
          type: string
          format: date-time
        pagination:
          $ref: "#/components/schemas/Pagination"
...
    ApiSuccessEnvelope:
      type: object
      additionalProperties: false
      required:
        - success
        - data
        - meta
        - error
      properties:
        success:
          type: boolean
          const: true
        data:
          type: object
          additionalProperties: true
        meta:
          $ref: "#/components/schemas/ApiMeta"
        error:
          type: "null"
...